[@ware]Modify ChangeToekn type from int to long

+++ nemakiware/src/main/java/jp/aegif/nemaki/service/dao/impl/ContentDaoServiceImpl.java
@@ -575,7 +575,7 @@ public class ContentDaoServiceImpl implements ContentDaoService {
 	}
 
 	@Override
-	public List<Change> getLatestChanges(int startToken, int maxItems) {
+	public List<Change> getLatestChanges(long startToken, int maxItems) {
 		return nonCachedContentDaoService
 				.getLatestChanges(startToken, maxItems);
 	}
